requirements.txt
djangorestframework == 3.15.2
Django == 5.1.3
mysqlclient == 2.2.6
